moves get sli handler to sli package
