+++ internal/event_handler/handler.go
@@ -6,6 +6,7 @@ import (
 	"github.com/keptn-contrib/dynatrace-service/internal/deployment"
 	"github.com/keptn-contrib/dynatrace-service/internal/monitoring"
 	"github.com/keptn-contrib/dynatrace-service/internal/problem"
+	"github.com/keptn-contrib/dynatrace-service/internal/sli"
 	keptnevents "github.com/keptn/go-utils/pkg/lib"
 	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
 	log "github.com/sirupsen/logrus"
@@ -32,7 +33,7 @@ func NewEventHandler(event cloudevents.Event) (DynatraceEventHandler, error) {
 	case keptnv2.GetFinishedEventType(keptnv2.ActionTaskName):
 		return &problem.ActionHandler{Event: event, DTConfigGetter: dtConfigGetter}, nil
 	case keptnv2.GetTriggeredEventType(keptnv2.GetSLITaskName):
-		return &GetSLIEventHandler{event: event, dtConfigGetter: dtConfigGetter}, nil
+		return &sli.GetSLIEventHandler{Event: event, DTConfigGetter: dtConfigGetter}, nil
 	default:
 		return deployment.NewCDEventHandler(event, dtConfigGetter), nil
 	}
